fixed stars on cards, consolidated outfitlist cards and related items cards into one card component
