+++ client/src/components/related_items/RelatedItemsList.jsx
@@ -1,10 +1,10 @@
 import React, {useState, useEffect} from 'react';
-import RelatedItemCard from './RelatedItemCard.jsx'
+import Card from './Card.jsx'
 import axios from 'axios';
 import Promise from 'bluebird';
 import Carousel from './Carousel.jsx'
 
-const RelatedItemsList = ({IDlist,  currProduct, handleRelatedItemClick}) => {
+const RelatedItemsList = ({IDlist, currProduct, handleRelatedItemClick}) => {
     const [list, setList] = useState ([])
     const [responses, setResponses] = useState([]);
     const [lastIndex, setLastIndex] = useState(5)
@@ -41,7 +41,11 @@ const RelatedItemsList = ({IDlist,  currProduct, handleRelatedItemClick}) => {
             <Carousel startIndex={startIndex} setStartIndex={setStartIndex} lastIndex={lastIndex} setLastIndex={setLastIndex} maxIndex={list.length}>
                 {list.map((card,index)=>{
                      if(index >= startIndex && index < lastIndex) 
-                        return <RelatedItemCard key={card.id} card={card} currProduct={ currProduct} handleRelatedItemClick={handleRelatedItemClick}/>
+                        return <Card key={card.id}  
+                        card={card} 
+                        currProduct={ currProduct} 
+                        handleRelatedItemClick={handleRelatedItemClick}
+                        relatedItems={true}/>
                 })}
                 {/* {list.map((card,index)=>{
                         return <RelatedItemCard key={card.id} card={card} currProduct={ currProduct} handleRelatedItemClick={handleRelatedItemClick}/>
